Move internal types to folder
