+++ src/Grpc.AspNetCore.Server.GrpcClient/Internal/GrpcHttpClientFactory.cs
@@ -24,10 +24,9 @@ using Grpc.Core;
 using Grpc.NetCore.HttpClient;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Http;
 using Microsoft.Extensions.Options;
 
-namespace Grpc.AspNetCore.Server.GrpcClient
+namespace Grpc.AspNetCore.Server.GrpcClient.Internal
 {
     internal class GrpcHttpClientFactory<TClient> : INamedTypedHttpClientFactory<TClient> where TClient : ClientBase
     {
